+++ system/modules/calendar/modules/ModuleCalendar.php
@@ -109,8 +109,6 @@ class ModuleCalendar extends \Events
 
 		// Find the boundaries
 		$objMinMax = \CalendarEventsModel::findBoundaries($this->cal_calendar);
-		$intLeftBoundary = date('Ym', $objMinMax->dateFrom);
-		$intRightBoundary = date('Ym', max($objMinMax->dateTo, $objMinMax->repeatUntil));
 
 		// Instantiate the template
 		$objTemplate = new \FrontendTemplate(($this->cal_ctemplate ? $this->cal_ctemplate : 'cal_default'));
@@ -128,7 +126,7 @@ class ModuleCalendar extends \Events
 		$intPrevYm = intval($prevYear . str_pad($prevMonth, 2, 0, STR_PAD_LEFT));
 
 		// Only generate a link if there are events (see #4160)
-		if ($intPrevYm >= $intLeftBoundary)
+		if ($objMinMax->dateFrom !== null && $intPrevYm >= date('Ym', $objMinMax->dateFrom))
 		{
 			$objTemplate->prevHref = $this->strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '?id=' . \Input::get('id') . '&amp;' : '?') . 'month=' . $intPrevYm;
 			$objTemplate->prevTitle = specialchars($lblPrevious);
@@ -146,7 +144,7 @@ class ModuleCalendar extends \Events
 		$intNextYm = $nextYear . str_pad($nextMonth, 2, 0, STR_PAD_LEFT);
 
 		// Only generate a link if there are events (see #4160)
-		if ($intNextYm <= $intRightBoundary)
+		if ($objMinMax->dateTo !== null && $intNextYm <= date('Ym', max($objMinMax->dateTo, $objMinMax->repeatUntil)))
 		{
 			$objTemplate->nextHref = $this->strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '?id=' . \Input::get('id') . '&amp;' : '?') . 'month=' . $intNextYm;
 			$objTemplate->nextTitle = specialchars($lblNext);