+++ system/modules/core/library/Contao/Controller.php
@@ -108,6 +108,12 @@ abstract class Controller extends \System
 			}
 		}
 
+		// Do not look for back end templates in theme folders (see #5379)
+		if ($strPrefix == 'be_' || $strPrefix == 'mail_')
+		{
+			return $arrTemplates;
+		}
+
 		// Try to select the themes (see #5210)
 		try
 		{
@@ -216,7 +222,7 @@ abstract class Controller extends \System
 
 			while ($objArticles->next())
 			{
-				$return .= $this->getArticle($objArticles, $blnMultiMode, false, $strColumn);
+				$return .= $this->getArticle($objArticles->current(), $blnMultiMode, false, $strColumn);
 			}
 
 			return $return;